+++ packages/core/src/components/modal/modal.tsx
@@ -1,4 +1,3 @@
-import { ModalRole } from "@dso-toolkit/sources";
 import { h, Component, ComponentInterface, Element, Event, EventEmitter, Host, Prop, State } from "@stencil/core";
 import { createFocusTrap, FocusTrap } from "focus-trap";
 import { v4 } from "uuid";
@@ -6,6 +5,8 @@ import { v4 } from "uuid";
 export interface DsoModalCloseEvent {
   originalEvent?: MouseEvent;
 }
+
+export type ModalRole = "alert" | "dialog" | "alertdialog";
 
 @Component({
   tag: "dso-modal",